+++ backend/app/project/tests/test_cli_commands.py
@@ -31,6 +31,8 @@ from project.api.models import (
     Contact,
     Device,
     DeviceContactRole,
+    ExportControl,
+    ManufacturerModel,
     Platform,
     PlatformContactRole,
     Site,
@@ -603,6 +605,213 @@ class TestCliCommands(BaseTestCase):
             reloaded_tsm_endpoint = db.session.get(TsmEndpoint, 1)
             self.assertEqual(reloaded_tsm_endpoint.name, "BAR")
             self.assertEqual(reloaded_tsm_endpoint.url, "https://bar")
+
+    def test_load_data_update_existing_by_unique_criteria(self):
+        """Ensure we can update existing data without knowing the primary key."""
+        with no_expire():
+            self.assertEqual(0, db.session.query(TsmEndpoint).count())
+            tsm_endpoint = TsmEndpoint(name="foo", url="https://foo")
+            db.session.add(tsm_endpoint)
+            db.session.commit()
+            data = [
+                {
+                    "fields": {
+                        "name": "foo",
+                        "url": "https://bar",
+                    },
+                    "model": "TsmEndpoint",
+                    "unique": ["name"],
+                },
+            ]
+            with tempfile.NamedTemporaryFile(mode="w+t") as temp_file:
+                json.dump(data, temp_file)
+                path = temp_file.name
+                temp_file.flush()
+                runner = CliRunner()
+                result = runner.invoke(
+                    loaddata,
+                    [path],
+                    env={"FLASK_APP": "manage"},
+                )
+            self.assertEqual(result.exit_code, 0)
+
+            self.assertEqual(1, db.session.query(TsmEndpoint).count())
+            reloaded_tsm_endpoint = (
+                db.session.query(TsmEndpoint).filter_by(name="foo").first()
+            )
+            self.assertEqual(reloaded_tsm_endpoint.url, "https://bar")
+
+    def test_load_data_insert_new_by_unique_criteria(self):
+        """Ensure we can insert new data without having a pk."""
+        with no_expire():
+            self.assertEqual(0, db.session.query(TsmEndpoint).count())
+            tsm_endpoint = TsmEndpoint(name="foo", url="https://foo")
+            db.session.add(tsm_endpoint)
+            db.session.commit()
+            data = [
+                {
+                    "fields": {
+                        "name": "bar",
+                        "url": "https://bar",
+                    },
+                    "model": "TsmEndpoint",
+                    "unique": ["name"],
+                },
+            ]
+            with tempfile.NamedTemporaryFile(mode="w+t") as temp_file:
+                json.dump(data, temp_file)
+                path = temp_file.name
+                temp_file.flush()
+                runner = CliRunner()
+                result = runner.invoke(
+                    loaddata,
+                    [path],
+                    env={"FLASK_APP": "manage"},
+                )
+            self.assertEqual(result.exit_code, 0)
+
+            self.assertEqual(2, db.session.query(TsmEndpoint).count())
+            reloaded_tsm_endpoint = (
+                db.session.query(TsmEndpoint).filter_by(name="bar").first()
+            )
+            self.assertEqual(reloaded_tsm_endpoint.url, "https://bar")
+
+    def test_load_data_error_if_unique_criteria_is_not_unique(self):
+        """Ensure don't change anything but raise an error if the unique criteria is not unique."""
+        with no_expire():
+            self.assertEqual(0, db.session.query(TsmEndpoint).count())
+            tsm_endpoint1 = TsmEndpoint(id=1, name="foo", url="https://foo1")
+            db.session.add(tsm_endpoint1)
+            tsm_endpoint2 = TsmEndpoint(id=2, name="foo", url="https://foo2")
+            db.session.add(tsm_endpoint2)
+            db.session.commit()
+            data = [
+                {
+                    "fields": {
+                        "name": "new",
+                        "url": "https://bar",
+                    },
+                    "model": "TsmEndpoint",
+                    "unique": ["name"],
+                },
+                {
+                    "fields": {
+                        "name": "foo",
+                        "url": "https://bar",
+                    },
+                    "model": "TsmEndpoint",
+                    "unique": ["name"],
+                },
+            ]
+            with tempfile.NamedTemporaryFile(mode="w+t") as temp_file:
+                json.dump(data, temp_file)
+                path = temp_file.name
+                temp_file.flush()
+                runner = CliRunner()
+                result = runner.invoke(
+                    loaddata,
+                    [path],
+                    env={"FLASK_APP": "manage"},
+                )
+            self.assertNotEqual(result.exit_code, 0)
+
+            # Nothing has changed.
+            self.assertEqual(2, db.session.query(TsmEndpoint).count())
+            reloaded_tsm_endpoint1 = db.session.get(TsmEndpoint, 1)
+            reloaded_tsm_endpoint2 = db.session.get(TsmEndpoint, 2)
+            self.assertEqual(reloaded_tsm_endpoint1.url, "https://foo1")
+            self.assertEqual(reloaded_tsm_endpoint2.url, "https://foo2")
+
+    def test_load_data_referencing_is_possible(self):
+        """Ensure we can use references between the entries."""
+        with no_expire():
+            self.assertEqual(0, db.session.query(ManufacturerModel).count())
+            self.assertEqual(0, db.session.query(ExportControl).count())
+            existing_manufacturer_model = ManufacturerModel(
+                manufacturer_name="TRUEBENER GmbH", model="SMT100"
+            )
+            db.session.add(existing_manufacturer_model)
+            existing_export_control = ExportControl(
+                manufacturer_model=existing_manufacturer_model, dual_use=False
+            )
+            db.session.add(existing_export_control)
+            db.session.commit()
+            data = [
+                {
+                    "model": "ManufacturerModel",
+                    "unique": ["manufacturer_name", "model"],
+                    "fields": {
+                        "manufacturer_name": "TRUEBENER GmbH",
+                        "model": "SMT100",
+                    },
+                    "result": "truebener_smt_100",
+                },
+                {
+                    "model": "ManufacturerModel",
+                    "unique": ["manufacturer_name", "model"],
+                    "fields": {
+                        "manufacturer_name": "GFZ",
+                        "model": "Custom",
+                    },
+                    "result": "gfz_custom",
+                },
+                {
+                    "model": "ExportControl",
+                    "unique": ["manufacturer_model"],
+                    "fields": {
+                        "dual_use": False,
+                        "internal_note": "no dual use according to manufacturer",
+                    },
+                    "references": {
+                        "manufacturer_model": "truebener_smt_100",
+                    },
+                },
+                {
+                    "model": "ExportControl",
+                    "unique": ["manufacturer_model"],
+                    "fields": {
+                        "internal_note": "Still unknown",
+                    },
+                    "references": {
+                        "manufacturer_model": "gfz_custom",
+                    },
+                },
+            ]
+            with tempfile.NamedTemporaryFile(mode="w+t") as temp_file:
+                json.dump(data, temp_file)
+                path = temp_file.name
+                temp_file.flush()
+                runner = CliRunner()
+                result = runner.invoke(
+                    loaddata,
+                    [path],
+                    env={"FLASK_APP": "manage"},
+                )
+
+            self.assertEqual(result.exit_code, 0)
+
+            reloaded_existing_export_control = (
+                db.session.query(ExportControl)
+                .filter_by(manufacturer_model_id=existing_manufacturer_model.id)
+                .first()
+            )
+
+            self.assertEqual(reloaded_existing_export_control.dual_use, False)
+            self.assertEqual(
+                reloaded_existing_export_control.internal_note,
+                "no dual use according to manufacturer",
+            )
+            reloaded_new_export_control = (
+                db.session.query(ExportControl)
+                .join(ManufacturerModel)
+                .filter(
+                    ManufacturerModel.manufacturer_name == "GFZ",
+                    ManufacturerModel.model == "Custom",
+                )
+                .first()
+            )
+            self.assertEqual(reloaded_new_export_control.dual_use, None)
+            self.assertEqual(reloaded_new_export_control.internal_note, "Still unknown")
 
     def test_load_data_empty_data(self):
         """Ensure we can run it without any data in it."""